pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "coq-modeling"
version = "0.0.1"
dependencies = [
    "torch",
    "openai",
    "transformers",
    "tqdm",
    "trl",
    "peft",
    "jsonlines",
    "sentencepiece",
    "protobuf",
    "matplotlib",
    "matplotlib-venn",
    "accelerate",
    "bitsandbytes",
    "scipy",
    "pyyaml",
    "pytest",
    "json-rpc",
    "hypothesis",
    "evaluate",
    "numpy",
    "termcolor",
    "types-PyYAML",
    "types-requests",
    "black",
    "GitPython",
    "ipdb",
    "parsy",
    #"edist",
		"Werkzeug",
    "waitress",
    "tiktoken",
]

[tool.setuptools.packages.find]
where = ["src", "test"]
